optionally find neighbors via sqlite3 + eve db
When the 'sqlite3' can be loaded and the file 'eveUniverseData.dx' is in
the directory of the script, arguments like "VFK-IV+3" are parsed as
"all systems within 3 jumps of VFK-IV" instead of as a partial system
name pattern.

On the upside, that requires less typing, on the downside for now it's
less resilient to people reporting hostiles with typoed variants of
actual system names. "i3o" and "ii-509", I'm looking at y'all.